Relocate plugins and utils module to common folder
