+++ src/app/main.js
@@ -1,4 +1,4 @@
-import './plugins';
+import '../common/plugins';
 import App from './application/app';
 
 const app = new App();
