+++ src/common/utils/global_channel.js
